add script for generating vectorizer
